requirements.txt
fastapi
uvicorn[standard]
boto3
pyjwt[crypto]
requests
pillow
matplotlib
numpy
cryptography
python-multipart